requirements.txt
anthropic>=0.40.0
openai>=1.50.0
ollama>=0.3.0
gradio>=4.44.0
python-dotenv>=1.0.0
pandas>=2.0.0
pydantic>=2.0.0
httpx>=0.27.0